Added loadMap function
